requirements.txt
bidict>=0.21.2, <1
certifi>=2020.12.5
chardet>=3.0.4, <4
click>=7.1.2, <8
colorama>=0.4.4, <1
dnspython>=1.16.0, <2
Flask>=1.1.2, <2
Flask-SocketIO>=5.0.1, <6
idna>=2.10, <3
itsdangerous>=1.1.0, <2
Jinja2>=2.11.3, <3
MarkupSafe>=1.1.1, <2
pyserial>=3.5, <4
python-engineio>=4.0.0, <5
python-socketio>=5.0.4, <6
requests>=2.25.1, <3
six>=1.15.0, <2
termcolor>=1.1.0, <2
urllib3>=1.26.3, <2
Werkzeug>=1.0.1, <2
zope.event>=4.5.0, <5
zope.interface>=5.2.0, <6
